should think about xdg sessions
Fixed the issue I had when trying to launch bspwm with plasma
installed, the plasma shell would appear inside bspwm, and it was caused
by "dex -a" which was launching everything in the autostart folder.

Should be able to make a bspwm empty service to launch others, bg timer
is running even on plasma and I don't need it there.